news-api/requirements.txt
Flask==3.0.3
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.4
openai==1.6.1
numpy==1.26.3
opencv-python-headless==4.9.0.80
pillow==10.2.0
uuid==1.30
pydub==0.25.1
opencv-python==4.9.0.80
moviepy==1.0.3
textwrap3==0.9.2
pydub==0.25.1
python-dotenv==1.0.1
aiohttp==3.10.10
pandas==2.2.3
ace-tools==0.0
Flash-Cors==5.0.0
bs4==0.0.2
mysql-connector-python==9.1.0